+++ lib/src/clixon_options.c
@@ -871,7 +871,7 @@ clicon_startup_mode(clicon_handle h)
     return clicon_str2int(startup_mode_map, mode);
 }
 
-/*! Which privileges drop method to use
+/*! Which privileges drop method to use for backend
  * @param[in] h     Clicon handle
  * @retval    mode  Privileges mode
  */
@@ -881,6 +881,20 @@ clicon_backend_privileges_mode(clicon_handle h)
     char *mode;
 
     if ((mode = clicon_option_str(h, "CLICON_BACKEND_PRIVILEGES")) == NULL)
+	return -1;
+    return clicon_str2int(priv_mode_map, mode);
+}
+
+/*! Which privileges drop method to use for restconf
+ * @param[in] h     Clicon handle
+ * @retval    mode  Privileges mode
+ */
+enum priv_mode_t
+clicon_restconf_privileges_mode(clicon_handle h)
+{
+    char *mode;
+
+    if ((mode = clicon_option_str(h, "CLICON_RESTCONF_PRIVILEGES")) == NULL)
 	return -1;
     return clicon_str2int(priv_mode_map, mode);
 }